play/multiagent/requirements.txt
anthropic
openai
pyyaml
